[Enhancement] add default hash function and dump query resource usage
