+++ be/src/exec/pipeline/exchange/sink_buffer.h
@@ -135,29 +135,30 @@ private:
     /// because TUniqueId::hi is exactly the same in one query
 
     // num eos per instance
-    phmap::flat_hash_map<int64_t, int64_t> _num_sinkers;
-    phmap::flat_hash_map<int64_t, int64_t> _request_seqs;
+    phmap::flat_hash_map<int64_t, int64_t, StdHash<int64_t>> _num_sinkers;
+    phmap::flat_hash_map<int64_t, int64_t, StdHash<int64_t>> _request_seqs;
     // Considering the following situation
     // Sending request 1, 2, 3 in order with one possible order of response 1, 3, 2,
     // and field transformation are as following
     //      a. receive response-1, _max_continuous_acked_seqs[x]->1, _discontinuous_acked_seqs[x]->()
     //      b. receive response-3, _max_continuous_acked_seqs[x]->1, _discontinuous_acked_seqs[x]->(3)
     //      c. receive response-2, _max_continuous_acked_seqs[x]->3, _discontinuous_acked_seqs[x]->()
-    phmap::flat_hash_map<int64_t, int64_t> _max_continuous_acked_seqs;
-    phmap::flat_hash_map<int64_t, std::unordered_set<int64_t>> _discontinuous_acked_seqs;
+    phmap::flat_hash_map<int64_t, int64_t, StdHash<int64_t>> _max_continuous_acked_seqs;
+    phmap::flat_hash_map<int64_t, std::unordered_set<int64_t>, StdHash<int64_t>> _discontinuous_acked_seqs;
     std::atomic<int32_t> _total_in_flight_rpc = 0;
     std::atomic<int32_t> _num_uncancelled_sinkers = 0;
     std::atomic<int32_t> _num_remaining_eos = 0;
 
     // The request needs the reference to the allocated finst id,
     // so cache finst id for each dest fragment instance.
-    phmap::flat_hash_map<int64_t, PUniqueId> _instance_id2finst_id;
-    phmap::flat_hash_map<int64_t, std::queue<TransmitChunkInfo, std::list<TransmitChunkInfo>>> _buffers;
-    phmap::flat_hash_map<int64_t, int32_t> _num_finished_rpcs;
-    phmap::flat_hash_map<int64_t, int32_t> _num_in_flight_rpcs;
-    phmap::flat_hash_map<int64_t, TimeTrace> _network_times;
-    phmap::flat_hash_map<int64_t, std::unique_ptr<Mutex>> _mutexes;
-    phmap::flat_hash_map<int64_t, TNetworkAddress> _dest_addrs;
+    phmap::flat_hash_map<int64_t, PUniqueId, StdHash<int64_t>> _instance_id2finst_id;
+    phmap::flat_hash_map<int64_t, std::queue<TransmitChunkInfo, std::list<TransmitChunkInfo>>, StdHash<int64_t>>
+            _buffers;
+    phmap::flat_hash_map<int64_t, int32_t, StdHash<int64_t>> _num_finished_rpcs;
+    phmap::flat_hash_map<int64_t, int32_t, StdHash<int64_t>> _num_in_flight_rpcs;
+    phmap::flat_hash_map<int64_t, TimeTrace, StdHash<int64_t>> _network_times;
+    phmap::flat_hash_map<int64_t, std::unique_ptr<Mutex>, StdHash<int64_t>> _mutexes;
+    phmap::flat_hash_map<int64_t, TNetworkAddress, StdHash<int64_t>> _dest_addrs;
 
     // True means that SinkBuffer needn't input chunk and send chunk anymore,
     // but there may be still in-flight RPC running.
